+++ arroyo/processing/strategies/batching.py
@@ -3,11 +3,14 @@ from __future__ import annotations
 import logging
 import time
 from abc import ABC, abstractmethod
+from collections import deque
 from dataclasses import dataclass, field
 from datetime import datetime
 from typing import (
     Callable,
+    Deque,
     Generic,
+    Iterator,
     Mapping,
     MutableMapping,
     MutableSequence,
@@ -17,10 +20,11 @@ from typing import (
 )
 
 from arroyo.processing.strategies.abstract import (
+    MessageRejected,
     ProcessingStrategy,
     ProcessingStrategyFactory,
 )
-from arroyo.types import Message, Partition, Position, TPayload
+from arroyo.types import Message, OffsetRange, Partition, Position, TPayload
 from arroyo.utils.metrics import get_metrics
 
 logger = logging.getLogger(__name__)
@@ -261,4 +265,230 @@ class BatchProcessingStrategyFactory(ProcessingStrategyFactory[TPayload]):
             self.__worker,
             self.__max_batch_size,
             self.__max_batch_time,
+        )
+
+
+@dataclass(frozen=True)
+class MessageBatch(Generic[TPayload]):
+    """
+    Represents a batch of messages with the range of offsets covered.
+    The range of offsets is represented as a mapping between partitions
+    and offset ranges.
+    The lowest offset in each offset range is inclusize. The highest is
+    esclusive.
+    """
+
+    messages: Sequence[Message[TPayload]]
+    offsets: Mapping[Partition, OffsetRange]
+
+    def __len__(self) -> int:
+        return len(self.messages)
+
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__}: {len(self)} message{'s' if len(self) != 1 else ''}>"
+
+    def __iter__(self) -> Iterator[Message[TPayload]]:
+        for m in self.messages:
+            yield m
+
+    def __getitem__(self, index: int) -> Message[TPayload]:
+        return self.messages[index]
+
+
+class OutOfOrderMessage(Exception):
+    pass
+
+
+class BatchBuilder(Generic[TPayload]):
+    """
+    Accumulates messages in a `MessageBatch` object.
+    It requires offsets to be in monotonic order in each partition.
+    """
+
+    def __init__(
+        self,
+        max_batch_time: float,
+        max_batch_size: int,
+    ) -> None:
+        self.__max_batch_time = max_batch_time
+        self.__max_batch_size = max_batch_size
+        self.__messages: MutableSequence[Message[TPayload]] = []
+        self.__offsets: MutableMapping[Partition, OffsetRange] = {}
+        self.__init_time = time.time()
+
+    def is_ready(self) -> bool:
+        return len(self.__messages) >= self.__max_batch_size or (
+            time.time() > self.__init_time + self.__max_batch_time
+        )
+
+    def append(self, message: Message[TPayload]) -> None:
+        if message.partition in self.__offsets:
+            if self.__offsets[message.partition].hi >= message.next_offset:
+                raise OutOfOrderMessage(
+                    f"Received offset {message.next_offset}. Current watermark {self.__offsets[message.partition].hi}"
+                )
+
+            self.__offsets[message.partition].hi = message.next_offset
+            self.__offsets[message.partition].timestamp = message.timestamp
+        else:
+            self.__offsets[message.partition] = OffsetRange(
+                message.offset, message.next_offset, message.timestamp
+            )
+
+        self.__messages.append(message)
+
+    def build(self) -> MessageBatch[TPayload]:
+        return MessageBatch(
+            messages=self.__messages,
+            offsets=self.__offsets,
+        )
+
+
+class BatchStep(ProcessingStrategy[TPayload]):
+    """
+    Accumulates messages into a batch. When the batch is full this
+    strategy submits it to the next step.
+
+    A batch is represented as a `MessageBatch` object, which includes
+    both the messages and the offset ranges.
+
+    A message is closed and submitted when the maximum number of messages
+    is received or when the max_batch_time has passed since the first
+    message was received.
+
+    This step does not allow out of order processing. The batch can only
+    be built with monotonically increasing offsets per partition.
+    If messages out of order are spot, this strategy will raise an
+    `OutOfOrderMessage` exception.
+    """
+
+    def __init__(
+        self,
+        max_batch_time_sec: float,
+        max_batch_size: int,
+        next_step: ProcessingStrategy[MessageBatch[TPayload]],
+    ) -> None:
+        self.__max_batch_time_sec = max_batch_time_sec
+        self.__max_batch_size = max_batch_size
+        self.__next_step = next_step
+        self.__batch_builder: Optional[BatchBuilder[TPayload]] = None
+        self.__closed = False
+
+    def __flush(self) -> None:
+        assert self.__batch_builder is not None
+        batch = self.__batch_builder.build()
+        if len(batch) > 0:
+            last_msg = batch.messages[-1]
+            self.__next_step.submit(
+                Message(
+                    partition=last_msg.partition,
+                    offset=last_msg.offset,
+                    timestamp=last_msg.timestamp,
+                    payload=batch,
+                )
+            )
+        self.__batch_builder = None
+
+    def submit(self, message: Message[TPayload]) -> None:
+        assert not self.__closed
+        if self.__batch_builder is None:
+            self.__batch_builder = BatchBuilder(
+                max_batch_time=self.__max_batch_time_sec,
+                max_batch_size=self.__max_batch_size,
+            )
+
+        self.__batch_builder.append(message)
+
+        if self.__batch_builder is not None and self.__batch_builder.is_ready():
+            try:
+                self.__flush()
+            except MessageRejected:
+                pass
+
+    def poll(self) -> None:
+        assert not self.__closed
+
+        if self.__batch_builder is not None and self.__batch_builder.is_ready():
+            try:
+                self.__flush()
+            except MessageRejected:
+                pass
+
+        self.__next_step.poll()
+
+    def close(self) -> None:
+        self.__closed = True
+
+    def terminate(self) -> None:
+        self.__closed = True
+        self.__batch_builder = None
+
+    def join(self, timeout: Optional[float] = None) -> None:
+        deadline = time.time() + timeout if timeout is not None else None
+        if self.__batch_builder is not None:
+            self.__flush()
+
+        self.__next_step.join(
+            timeout=max(deadline - time.time(), 0) if deadline is not None else None
+        )
+
+
+class UnbatchStep(ProcessingStrategy[MessageBatch[TPayload]]):
+    """
+    This processing step receives batches and explodes them sending
+    the content message by message to the next step.
+
+    A batch is represented as a `MessageBatch` object.
+
+    If this step receives a `MessageRejected` exception from the next
+    step it would keep the remaining messages and attempt to submit
+    them at the following call to `poll`
+    """
+
+    def __init__(
+        self,
+        next_step: ProcessingStrategy[TPayload],
+    ) -> None:
+        self.__next_step = next_step
+        self.__batch_to_send: Deque[Message[TPayload]] = deque()
+        self.__closed = False
+
+    def __flush(self) -> None:
+        while self.__batch_to_send:
+            msg = self.__batch_to_send[0]
+            self.__next_step.submit(msg)
+            self.__batch_to_send.popleft()
+
+    def submit(self, message: Message[MessageBatch[TPayload]]) -> None:
+        assert not self.__closed
+
+        if self.__batch_to_send:
+            raise MessageRejected
+
+        self.__batch_to_send.extend(message.payload.messages)
+        self.__flush()
+
+    def poll(self) -> None:
+        assert not self.__closed
+
+        try:
+            self.__flush()
+        except MessageRejected:
+            pass
+
+        self.__next_step.poll()
+
+    def close(self) -> None:
+        self.__closed = True
+
+    def terminate(self) -> None:
+        self.__closed = True
+        self.__current_batch = None
+
+    def join(self, timeout: Optional[float] = None) -> None:
+        deadline = time.time() + timeout if timeout is not None else None
+        self.__flush()
+
+        self.__next_step.join(
+            timeout=max(deadline - time.time(), 0) if deadline is not None else None
         )
